campus-os-pre/server/requirements.txt
fastapi
uvicorn
sqlalchemy
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
requests
python-dotenv
